backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
pydantic
python-dotenv
openrouteservice